Fixed iTC Readout, Moved Zain's files to backup folder
Moved Zain's files to backup folder

Also fixed the iTC read/write error that was being caused by the GPIB
connecting timing out. The problem was in LabView, with how the
Instrument I/O Assistant handled (or rather, didn't handle) errors.
Switched the communication method from GPIB Code Generation to VISA Code
Generation, which let the I/O Ass. throw an error. At first, this would
just stop the program to alert the user to an error. It's now handled
internally through LabView, which will see if an error has occurred and
if so, won't display the error results on the strip chart, but will show
a red Error light and the most recent error message and records the
error into an external error file.

Minor other fixes were also implemented, such as correctly outputting
the date/time and column headings into the output file to make plotting
results easier.

Also inserted a timer directly in the GPIB flow so that it will run
once, wait x seconds, then run again.  This is different from the on/off
loop from Zain's work, which would turn communications on and do as many
as it could before the flow turned off again. This way we definitively
have only 1 measurement per x seconds.